% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{as.curve_set}
\alias{as.curve_set}
\title{Convert an envelope or fdata object to a curve_set object}
\usage{
as.curve_set(curve_set, ...)
}
\arguments{
\item{curve_set}{An object to be converted to a \code{\link{curve_set}} object.
The \code{envelope} objects of \pkg{spatstat} and \code{fdata} objects of
\pkg{fda.usc} are supported currently, besides curve_set objects.}

\item{...}{Allows passing arguments to \code{\link{create_curve_set}}.}
}
\value{
If an \code{envelope} object of \pkg{spatstat} or an
\code{\link[fda.usc]{fdata}} object is given, return a corresponding
curve_set object. If a curve_set object was given, check it and return it
unharmed.
}
\description{
If given an envelope object of \pkg{spatstat} or a \code{fdata} object of
\pkg{fda.usc}, convert it into a curve_set object. If given a curve_set
object, check its correctness and give it back.
}
