% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{check_curve_set_dimensions}
\alias{check_curve_set_dimensions}
\title{Check curve set dimensions}
\usage{
check_curve_set_dimensions(x)
}
\arguments{
\item{x}{A list of curve sets or \code{\link[spatstat]{envelope}} objects.}
}
\description{
Check curve set dimensions
}
