% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation.r
\name{check_deviation_set}
\alias{check_deviation_set}
\title{Check the object.}
\usage{
check_deviation_set(deviation_set)
}
\arguments{
\item{deviation_set}{A potential deviation_set object.}
}
\description{
Check the object.
}

