\name{check_deviation_set_content}
\alias{check_deviation_set_content}
\title{Check the content validity of a potential deviation_set object.}
\usage{
  check_deviation_set_content(deviation_set)
}
\arguments{
  \item{deviation_set}{A potential deviation_set object.}
}
\description{
  Check the content validity of a potential deviation_set
  object.
}

