% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaling.r
\name{check_probs}
\alias{check_probs}
\title{Check for an increasing two-element vector of probabilities.}
\usage{
check_probs(probs)
}
\arguments{
\item{probs}{A vector to be checked.}
}
\description{
Check for an increasing two-element vector of probabilities.
}

