% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/curve-set.r
\name{check_residualness}
\alias{check_residualness}
\title{Check that the curve_set object portrays residual curves.}
\usage{
check_residualness(curve_set)
}
\arguments{
\item{curve_set}{A curve_set or an \code{\link[spatstat]{envelope}}
object. If an envelope object is given, it must contain the summary
functions from the simulated patterns which can be achieved by setting
savefuns = TRUE when calling envelope().}
}
\description{
Check that the curve_set object portrays residual curves.
}
