% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{combine_curve_sets}
\alias{combine_curve_sets}
\title{Combine curve sets.}
\usage{
combine_curve_sets(x)
}
\arguments{
\item{x}{A list of curve sets or \code{\link[spatstat]{envelope}} objects.}
}
\value{
A curve set that is a combination of the curve sets given in 'x'.
}
\description{
Combine curve sets to a one curve set, e.g. for testing by means of several test functions.
}
