% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_envelope_tests.R
\name{combined_scaled_MAD_bounding_curves}
\alias{combined_scaled_MAD_bounding_curves}
\title{Calculate the lower and upper bounding curves of a combined global scaled MAD envelope test}
\usage{
combined_scaled_MAD_bounding_curves(central_curves_ls, max_u, lower_f, upper_f)
}
\arguments{
\item{central_curves_ls}{A list containing the central functions for different test functions.}

\item{max_u}{The k_alpha'th largest value of the u_i, i=1,...,nsim+1 for each individual test.}

\item{lower_f}{The first component in the object returned by \code{\link{combined_scaled_MAD_envelopefuns}}.}

\item{upper_f}{The second component in the object returned by \code{\link{combined_scaled_MAD_envelopefuns}}.}
}
\description{
Calculate the lower and upper bounding curves of a combined global scaled MAD envelope test
}

