% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_envelope_tests.R
\name{combined_scaled_MAD_bounding_curves_chars}
\alias{combined_scaled_MAD_bounding_curves_chars}
\title{Quantiles (qdir) or sds (st) for the combined scaled MAD envelope tests}
\usage{
combined_scaled_MAD_bounding_curves_chars(curve_sets, test = c("qdir", "st"),
  probs = c(0.025, 0.975))
}
\arguments{
\item{curve_sets}{A list of objects of type \code{\link{curve_set}} or \code{\link[spatstat]{envelope}}.}

\item{test}{Either "qdir" for the \code{\link{qdir_envelope}} test or
"st" for the \code{\link{st_envelope}} test.}

\item{probs}{A two-element vector containing the lower and upper
quantiles for the envelope, in that order and on the interval [0, 1].
The default values are 0.025 and 0.975.}
}
\description{
Quantiles (qdir) or sds (st) for the combined scaled MAD envelope tests, i.e.
the quantities are calculated for all the curve sets provided in the list "curve_sets"
and the results are returned in corresponding lists.
}

