\name{convert_envelope}
\alias{convert_envelope}
\title{Convert an envelope object to a curve_set object.}
\usage{
  convert_envelope(curve_set)
}
\description{
  If given an envelope object, convert it into a curve_set
  object. If given a curve_set object, check its
  correctness and give it back.
}

