% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.R
\name{curve_set_check_r}
\alias{curve_set_check_r}
\title{Check r values of a \code{\link{curve_set}} object for plotting}
\usage{
curve_set_check_r(x)
}
\arguments{
\item{x}{A curve_set or envelope_test object.}
}
\description{
Check r values of a \code{\link{curve_set}} object to find out if there is
one or several test functions. Find out breaking r values for plotting.
}

