% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deviation.r
\name{deviation}
\alias{deviation}
\title{Deviation of curves.}
\usage{
deviation(curve_set, measure = "max", ...)
}
\arguments{
\item{curve_set}{A residual curve_set object. Can be obtained by using
residual().}

\item{measure}{The deviation measure to use. Default is 
'max'. Must be one of the following: 'max', 'int' or 
'int2'.}

\item{...}{Arguments to be passed to the measure function, if applicable.}
}
\value{
A deviation_set object. The list has two elements: obs and sim.
  obs is scalar while sim is a vector with at least one element.
}
\description{
Deviation of curves.
}

