% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_envelopes.r
\name{dg.global_envelope_test}
\alias{dg.global_envelope_test}
\title{Adjusted global envelope tests}
\usage{
dg.global_envelope_test(X, nsim = 499, nsimsub = nsim, simfun = NULL,
  fitfun = NULL, simfun.arg = NULL, testfuns = NULL, ...,
  type = "erl", alpha = 0.05, alternative = c("two.sided", "less",
  "greater"), r_min = NULL, r_max = NULL, take_residual = FALSE,
  save.cons.envelope = savefuns || savepatterns, savefuns = FALSE,
  savepatterns = FALSE, verbose = TRUE, MrkvickaEtal2017 = FALSE,
  mc.cores = 1L)
}
\arguments{
\item{X}{An object containing point pattern data. A point pattern (object of class "ppp")
or a fitted point process model (object of class "ppm" or "kppm"). See
\code{\link[spatstat]{envelope}}.}

\item{nsim}{The number of simulations to be generated in the primary test.}

\item{nsimsub}{Number of simulations in each basic test. There will be nsim repetitions of the
basic test, each involving nsimsub simulated realisations, so there will be a total of
nsim * (1 + nsimsub) simulations.}

\item{simfun}{A function for generating simulations from the null model. If given, this function
is called by replicate(n=nsim, simfun(simfun.param), simplify=FALSE) to make nsim simulations.
The function should return an \code{\link[spatstat]{ppp}} object as those are further passed to
\code{\link[spatstat]{envelope}}.
If the function is not provided, then \code{\link[spatstat]{envelope}} is used also for generating
the point patterns from the null hypothesis.}

\item{fitfun}{A function for estimating the parameters of the null model. If not given, then
\code{\link[spatstat]{envelope}} takes care of the parameter estimation as well (and X should be a fitted
model object). The function 'fitfun' should return the fitted model in the form that it can be directly
passed to 'simfun' as the argument 'simfun.arg'.}

\item{simfun.arg}{The parameter to be passed to simfun. The function simfun should handle
with the structure of simfun.param. FIXME should be here or not?}

\item{testfuns}{A list of lists of parameters to be passed to \code{\link[spatstat]{envelope}}.
A list of parameters should be provided for each test function that is to be used in the combined test.}

\item{...}{Additional parameters to \code{\link[spatstat]{envelope}} in the case where only one test
function is used. In that case, this is an alternative to providing the parameter in the argument
testfuns. If \code{\link[spatstat]{envelope}} is also used to generate simulations under the null
hypothesis (if simfun not provided), then also recall to specify how to generate the simulations.}

\item{type}{The type of the global envelope with current options for 'rank', 'erl', 'cont', 'area',
'qdir', 'st' and 'unscaled'. See details.}

\item{alpha}{The significance level. The 100(1-alpha)\% global envelope will be calculated.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of the following: "two.sided" (default), "less" or "greater".
The last two options only available for types \code{'rank'}, \code{'erl'},
\code{'cont'} and \code{'area'}.}

\item{r_min}{The minimum radius to include in the test.}

\item{r_max}{The maximum radius to include in the test. Note: cannot be larger than r-values used
in calculating functions by \code{\link[spatstat]{envelope}}.}

\item{take_residual}{If (needed for visual reasons only) the theoretical or mean behaviour of the
test function is reduced from the test functions. If TRUE, then: If \code{\link[spatstat]{envelope}}
provides the theoretical value 'theo' for the simulated model, then this value is used. Otherwise,
the theoretical function is taken as the mean of the simulations.}

\item{save.cons.envelope}{Logical flag indicating whether to save the unadjusted envelope test results.}

\item{savefuns}{Logical flag indicating whether to save all the simulated function values.
See \code{\link[spatstat]{envelope}}.}

\item{savepatterns}{Logical flag indicating whether to save all the simulated point patterns.
See \code{\link[spatstat]{envelope}}.}

\item{verbose}{Logical flag indicating whether to print progress reports during the simulations.
See \code{\link[spatstat]{envelope}}.}

\item{MrkvickaEtal2017}{Logical. If TRUE, type is "st" or "qdir" and several test functions are used,
then the combined scaled MAD envelope presented in Mrkvička et al. (2017) is calculated. Otherwise,
the two-step procedure described in \code{\link{global_envelope_test}} is used for combining the tests.
Default to FALSE.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously.
Must be at least one, and parallelization requires at least two cores. On a Windows computer mc.cores must be 1
(no parallelization). For details, see \code{\link[parallel]{mclapply}}, for which the argument is passed.}
}
\value{
An object of class \code{global_envelope} or \code{combined_global_envelope}.
}
\description{
Adjusted global envelope tests.
}
\details{
The specification of X is important here:

1) If simfun = NULL and fitfun = NULL (default), then \code{\link[spatstat]{envelope}}
is used for generating simulations under the null hypothesis and
\itemize{
\item If X is a point pattern, the null hypothesis is CSR.
\item If X is a fitted model, the null hypothesis is that model.
}

2) The user can provide the function for fitting the model (fitfun) and for simulating
from the fitted model (simfun). These functions should be coupled with each other such
that the object returned by 'fitfun' is directly accepted as the (single) argument in 'simfun'.
Further X should then be an \code{\link[spatstat]{ppp}} object and 'fitfun' should accept as
the argument an \code{\link[spatstat]{ppp}} object (X and further simulated point patterns).

A note: The structure of the code, which utilizes \code{\link[spatstat]{envelope}} though an
inner function, mimics the structure in the function \code{\link[spatstat]{dg.envelope}}
in the use of \code{\link[spatstat]{envelope}}. However, this function allows for more general
use as described above and for the different global envelope tests.

For the rank envelope test, the test is the test described in Myllymäki et al. (2017) with the
adjustment of Baddeley et al. (2017).
For other test types, the test (also) uses the two-stage procedure of Dao and Genton (2014) with
the adjustment of Baddeley et al. (2017).

See examples in \code{\link{saplings}}.
}
\references{
Myllymäki, M., Mrkvička, T., Grabarnik, P., Seijo, H. and Hahn, U. (2017). Global envelope tests for spatial point patterns. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 79: 381-404. doi: 10.1111/rssb.12172

Mrkvička, T., Myllymäki, M. and Hahn, U. (2017) Multiple Monte Carlo testing, with applications in spatial point processes. Statistics & Computing 27(5): 1239-1255. DOI: 10.1007/s11222-016-9683-9

Dao, N.A. and Genton, M. (2014). A Monte Carlo adjusted goodness-of-fit test for parametric models describing spatial point patterns. Journal of Graphical and Computational Statistics 23, 497-517.

Baddeley, A., Hardegen, A., Lawrence, T., Milne, R. K., Nair, G. and Rakshit, S. (2017). On two-stage Monte Carlo tests of composite hypotheses. Computational Statistics and Data Analysis 114: 75-87. doi: http://dx.doi.org/10.1016/j.csda.2017.04.003
}
\seealso{
\code{\link{rank_envelope}}, \code{\link{qdir_envelope}}, \code{\link{st_envelope}},
\code{\link{plot.global_envelope}}, \code{\link{saplings}}
}
