% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.R
\name{env_basic_plot}
\alias{env_basic_plot}
\title{An internal spptest function for making a basic "global envelope plot".}
\usage{
env_basic_plot(x, main, ylim, xlab, ylab, color_outside,
  separate_yaxes = FALSE, max_ncols_of_plots = 2, ...)
}
\arguments{
\item{x}{An 'envelope_test' object.}

\item{main}{See \code{\link{plot.default}}.}

\item{ylim}{See \code{\link{plot.default}}.}

\item{xlab}{See \code{\link{plot.default}}.}

\item{ylab}{See \code{\link{plot.default}}.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used.}

\item{separate_yaxes}{Logical (default FALSE). By default also the combined envelope plots have
a common y-axis. If TRUE, then separate y-axes are used for different parts of a combined test.}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2.}

\item{...}{Additional parameters to be passed to the function \code{\link{plot}}.}
}
\description{
An internal spptest function for making a basic "global envelope plot".
}

