% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.r
\name{env_dotplot}
\alias{env_dotplot}
\title{An internal GET function for making a dotplot style "global envelope plot".}
\usage{
env_dotplot(x, main, ylim, xlab, ylab, color_outside = TRUE, labels, ...)
}
\arguments{
\item{x}{An 'envelope_test' object.}

\item{main}{See \code{\link{plot.default}}.}

\item{ylim}{See \code{\link{plot.default}}.}

\item{xlab}{See \code{\link{plot.default}}.}

\item{ylab}{See \code{\link{plot.default}}.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used.}

\item{labels}{Labels for the tests at x-axis.}

\item{...}{Additional parameters to be passed to the function \code{\link{plot}}.}
}
\description{
An internal GET function for making a dotplot style "global envelope plot".
}
