% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.r
\name{env_ggplot}
\alias{env_ggplot}
\title{An internal GET function for making a ggplot2 style "global envelope plot".}
\usage{
env_ggplot(x, base_size, main, ylim, xlab, ylab, separate_yaxes = FALSE,
  max_ncols_of_plots = 2, labels = NULL)
}
\arguments{
\item{x}{An 'envelope_test' object.}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_type="ggplot2"}.}

\item{main}{See \code{\link{plot.default}}.}

\item{ylim}{See \code{\link{plot.default}}.}

\item{xlab}{See \code{\link{plot.default}}.}

\item{ylab}{See \code{\link{plot.default}}.}

\item{separate_yaxes}{Logical (default FALSE). By default also the combined envelope plots have
a common y-axis. If TRUE, then separate y-axes are used for different parts of a combined test.}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2.}

\item{labels}{Labels for the separate plots. Ignored if separate_yaxes is FALSE.}
}
\description{
An internal GET function for making a ggplot2 style "global envelope plot".
}
