% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.R
\name{env_main_default}
\alias{env_main_default}
\title{An internal spptest function for setting the default main for a global envelope plot.}
\usage{
env_main_default(x)
}
\arguments{
\item{x}{An 'envelope_test' object.}
}
\description{
An internal spptest function for setting the default main for a global envelope plot.
}

