% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envelope_helper.R
\name{env_ylim_default}
\alias{env_ylim_default}
\title{An internal GET function for setting the default ylim for a global envelope plot.}
\usage{
env_ylim_default(x, use_ggplot2)
}
\arguments{
\item{x}{An 'envelope_test' object.}

\item{use_ggplot2}{TRUE/FALSE, If TRUE, then default ylim are for \code{\link{env_ggplot}}.
Otherwise the ylim are for \code{\link{env_basic_plot}}.}
}
\description{
An internal GET function for setting the default ylim for a global envelope plot.
}

