% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/curve-set.r
\name{envelope_to_curve_set}
\alias{envelope_to_curve_set}
\title{Turn an \code{\link[spatstat]{envelope}} object into a curve_set object.}
\usage{
envelope_to_curve_set(env)
}
\arguments{
\item{env}{An \code{\link[spatstat]{envelope}} object. The envelope()
functions must have been called with savefuns = TRUE.}
}
\value{
A corresponding curve_set object.
}
\description{
Turn an \code{\link[spatstat]{envelope}} object into a curve_set object.
}
