% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-func.r
\name{estimate_K_f}
\alias{estimate_K_f}
\title{Estimate K_f.}
\usage{
estimate_K_f(pattern, mtf_name = "m", r_max = NULL, r_vec = NULL,
  calc_unmarked = TRUE, edge_correction = "translate",
  use_biased_lambda2 = FALSE)
}
\arguments{
\item{pattern}{A \code{\link[spatstat]{ppp}} object as the simple marked
point pattern to be analysed. The marks need to be in the form of a
numeric vector. The window has to have the type "rectangle".}

\item{mtf_name}{A vector of mark test function names. "1" stands for the
unmarked K-function. Accepted values are '1', 'm', 'mm', 'gamma',
'gammaAbs' and 'morAbs'.}

\item{r_max}{A positive scalar value representing the maximum radius that
should be considered. r_vec overrides r_max. By default, r_max is NULL
and will get a sensible default.}

\item{r_vec}{A monotonically increasing vector of non-negative r-values
to act as the endpoints of the bins for the K_f-functions. r_vec
overrides r_max. The bins are exclusive on the left and inclusive on
the right. If the first vector element has value zero, it will be
regarded as the collapsed bin [0, 0], and the next bin will start from
and exclude 0.}

\item{calc_unmarked}{Whether to include the unmarked K in the result.}

\item{edge_correction}{The name of the edge correction to be used. Options
are 'translate' and 'none'.}

\item{use_biased_lambda2}{A logical scalar on whether to use the biased
or the unbiased (in the Poisson case) estimate of the intensity
squared.}
}
\value{
A list with either two or three components. 'obs' has the summary
  function that was asked for. 'r' contains the radius values. 'unmarked'
  contains the unmarked K function, if asked for.
}
\description{
Estimate K_f.
}
