\name{estimate_p_value.deviation_set}
\alias{estimate_p_value.deviation_set}
\title{Estimate p-value.}
\usage{
  estimate_p_value.deviation_set(deviation_set, ...)
}
\arguments{
  \item{deviation_set}{A deviation_set object. Use
  deviation() to get one.}
}
\value{
  The p-value.
}
\description{
  Check \code{\link{estimate_p_value.default}} for details.
}

