% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p-value.r
\name{estimate_p_value}
\alias{estimate_p_value}
\title{Estimate p-value.}
\usage{
estimate_p_value(x, ...)
}
\arguments{
\item{x}{The first argument.}

\item{...}{Additional arguments.}
}
\description{
Estimate p-value.
}
\seealso{
\code{\link{estimate_p_value.default}}
}
