% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.r
\name{fdr_ATS}
\alias{fdr_ATS}
\title{Adaptive two stage method for FDR}
\usage{
fdr_ATS(
  curve_set,
  pvalues,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{curve_set}{A \code{curve_set} (see \code{\link{create_curve_set}}) or an
\code{envelope} object of \pkg{spatstat} containing the observed function
and the resampled functions from which the envelope is to be constructed.
If provided, resampling p-values will be calculated and used (pvalues then ignored).}

\item{pvalues}{A vector of p-values. An alternative input of curve_set.}

\item{alpha}{The significance level. The 100(1-alpha)\% global envelope will be calculated.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of the following: "two.sided" (default), "less" or "greater".
The last two options only available for types \code{'rank'}, \code{'erl'},
\code{'cont'} and \code{'area'}.}
}
\description{
Adaptive two stage method for FDR
}
\examples{
if(require("spatstat", quietly=TRUE)) {
  # CSR test
  #=========
  X <- unmark(spruces)
  # Number of simulations should be larger than
  alpha <- 0.10
  1 / (alpha/(1+alpha) / 513) # 513 is the number of r-values
  \donttest{nsim <- 6000 # Number of simulations}
  \dontshow{nsim <- 20}
  env <- envelope(X, fun="Lest", nsim=nsim,
                  savefuns=TRUE, # save the functions
                  correction="translate", # edge correction for L
                  transform = expression(.-r), # centering
                  simulate=expression(runifpoint(ex=X))) # Simulate CSR
  res1 <- fdr_ATS(env, alpha=alpha)
  plot(res1$pvalues, col=res1$rej*1+1) # red = rejected
}
}
