% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.r
\name{fdr_envelope}
\alias{fdr_envelope}
\title{The FDR envelope}
\usage{
fdr_envelope(
  curve_sets,
  alpha = 0.05,
  curve_sets2 = NULL,
  alternative = c("two.sided", "less", "greater"),
  algorithm = c("IATSE", "ATSE"),
  FDRest = c("ST", "YB", "YBu"),
  beta = 0.05
)
}
\arguments{
\item{curve_sets}{A \code{curve_set} (see \code{\link{create_curve_set}}) or an
\code{envelope} object of \pkg{spatstat} containing the observed function
and the functions from which the envelope is to be constructed.
Alternatively, a list of appropriate objects can be given.}

\item{alpha}{The significance level. The 100(1-alpha)\% global envelope will be calculated.
If a vector of values is provided, the global envelopes are calculated for each value.}

\item{curve_sets2}{A \code{curve_set} giving another set of simulations from which
the numbers of rejections under the null are calculated. (Alternatively, a list similarly
as \code{curve_sets}.) If not provided, then expectation under H0 is used instead following
Storey (2002) and Storey and Tibshirani (2003).}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of the following: "two.sided" (default), "less" or "greater".
The last two options only available for types \code{'rank'}, \code{'erl'},
\code{'cont'} and \code{'area'}.}

\item{algorithm}{Either "IATSE" or "ATSE" standing for the iteratively adaptive two-stage
envelope and the adaptive two-stage envelope, respectively, see Mrkvicka and Myllym채ki (2022).}

\item{FDRest}{Different options to estimate FDR. Recommended choice is ST.
The options YB and YBu were found to lead to liberal tests in Mrkvicka and Myllym채ki (2022);
they are provided for the experimenting purposes only.}

\item{beta}{A quantile for YBu method. Default to 0.05.}
}
\description{
Calculate the FDR envelope based on ATSE or IATSE algorithm.
}
\examples{
# A GLM example
data(rimov)
\donttest{nsim <- 1000 # Number of simulations}
\dontshow{nsim <- 20}
sims1 <- graph.flm(nsim=nsim,
                 formula.full = Y~Year,
                 formula.reduced = Y~1,
                 curve_sets = list(Y=rimov), factors = data.frame(Year = 1979:2014),
                 savefuns = TRUE)
res <- fdr_envelope(attr(sims1, "simfuns")[[1]])
plot(res)

}
\references{
Mrkvi훾ka and Myllym채ki (2022). False discovery rate envelopes. arXiv:2008.10108 [stat.ME]
}
