% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfanova.r
\name{frank.fanova}
\alias{frank.fanova}
\title{A functional ANOVA based on the rank envelope applied to F values}
\usage{
frank.fanova(nsim, x, groups, alpha = 0.05)
}
\arguments{
\item{nsim}{The number of random permutations.}

\item{x}{The original data (an array of functions). Typically a matrix or a data frame,
also \code{\link[fda.usc]{fdata}} objects allowed.}

\item{groups}{The original froups (a factor vector representing the assignment to groups).}

\item{alpha}{The significance level of the test.}
}
\description{
A functional ANOVA based on the rank envelope applied to F values
}
\examples{
data(rimov)
groups <- factor(c(rep(1, times=12), rep(2, times=12), rep(3, times=12)))
res <- frank.fanova(nsim=2499, x=rimov, groups=groups)
plot(res)
}
