% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_central_region.R
\docType{data}
\name{StatCentralRegion}
\alias{StatCentralRegion}
\alias{GeomCentralRegion}
\alias{GeomCentralRegionMulti}
\title{Central region plot}
\format{
An object of class \code{StatCentralRegion} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomCentralRegion} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomCentralRegionMulti} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatCentralRegion

GeomCentralRegion

GeomCentralRegionMulti
}
\description{
\code{ggproto} objects for central region plot. Not to be used directly.
}
\seealso{
\code{\link{geom_central_region}}
}
\keyword{datasets}
