% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-value.r
\name{p_value_ties_default}
\alias{p_value_ties_default}
\title{The default ties method for the p-value}
\usage{
p_value_ties_default()
}
\description{
The default ties method for the p-value calculated by estimate_p_value
}

