% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_envelopes.r
\name{plot.adjusted_combined_envelope_test}
\alias{plot.adjusted_combined_envelope_test}
\title{Plot method for the class 'adjusted_combined_envelope_test'}
\usage{
\method{plot}{adjusted_combined_envelope_test}(x, main, plot_type = c("rank", "MAD"), ...)
}
\arguments{
\item{x}{an 'adjusted_combined_envelope_test' object}

\item{main}{See \code{\link{plot.default}}. Default is x$method.}

\item{plot_type}{"rank" for the result of the rank envelope test; "MAD" for the
adjusted combined scaled MAD envelope. The latter only available if saved in 'x'.}

\item{...}{Additional parameters to be passed to \code{\link{plot.global_envelope}},
if plot_type is "rank" or to \code{\link{plot.combined_global_envelope}}, if
plot_type is "MAD".}
}
\description{
Plot method for the class 'adjusted_combined_envelope_test'
}
\seealso{
\code{\link{plot.global_envelope}}, \code{\link{plot.combined_global_envelope}}.
}
