% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.combined_global_envelope}
\alias{plot.combined_global_envelope}
\title{Plot method for the class 'combined_global_envelope'}
\usage{
\method{plot}{combined_global_envelope}(x, plot_style="basic", level = 1,
  main, ylim, xlab, ylab, use_ggplot2,
  separate_yaxes = FALSE, max_ncols_of_plots = 2,
  labels = NULL,
  color_outside = TRUE, env.col = 1, base_size = 15,
  add = FALSE, ...)
}
\arguments{
\item{x}{an 'combined_global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{level}{1 for plotting the combined global envelopes or
2 for plotting the second level ERL test result.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{use_ggplot2}{Logical, whether plot_style is "ggplot2" or not. Outdated, use the argument plot_style instead.}

\item{separate_yaxes}{Logical (default FALSE). If TRUE, then separate y-axes are used for
different parts of a combined test. If FALSE, then the components are displayed one next to each other
with a joint y-axis.}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2.}

\item{labels}{Labels for the separate plots (for \code{plot_style = "ggplot2"}).
Ignored if separate_yaxes is FALSE. Or, for \code{dotplot=TRUE}, labels for the tests at x-axis.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used. Relevant only for \code{plot_style = "basic"}.}

\item{env.col}{The color for the envelope lines (or dotplot arrows). Default 1 (black).}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{add}{Whether to add the plot to an existing plot (TRUE) or to draw a new plot (FALSE).}

\item{...}{Additional parameters to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'combined_global_envelope'
}
