% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes2d.r
\name{plot.combined_global_envelope_2d}
\alias{plot.combined_global_envelope_2d}
\title{Plot method for the class 'combined_global_envelope_2d'}
\usage{
\method{plot}{combined_global_envelope_2d}(x, sign.col = c(255, 0, 0), transparency = 85, ...)
}
\arguments{
\item{x}{an 'combined_global_envelope_2d' object}

\item{sign.col}{A vector of length 3 giving the color for the significant regions.
The elements should correspond to red, green, blue of \code{\link[grDevices]{rgb}}.
Default to red (255, 0, 0).}

\item{transparency}{A number between 0 and 255 (default 85, 33% transparency).
Corresponds to alpha of \code{\link[grDevices]{rgb}}. Used in plotting the significant regions.}

\item{...}{Additional parameters to be passed to \code{\link[spatstat]{plot.im}}.}
}
\description{
Plot method for the class 'combined_global_envelope_2d'
}
\details{
Additional parameter \code{col} can be passed in \code{...} to \code{\link[spatstat]{plot.im}}.
If \code{col} not given, a \code{\link[spatstat]{colourmap}} of 255 grey values between the
minimum and maximum of the function values is used for each image separately.
If \code{col} is provided, the same specification will be used for each produced plot,
which may make it easier to compare the figures with each other.
}
