% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_envelope_tests.R
\name{plot.combined_scaled_MAD_test}
\alias{plot.combined_scaled_MAD_test}
\title{Plot method for the class 'combined_scaled_MAD_test'}
\usage{
\method{plot}{combined_scaled_MAD_test}(x, plot_type = c("envelope", "rank"), ...)
}
\arguments{
\item{x}{An 'combined_scaled_MAD_test' object}

\item{plot_type}{"rank" or "qdir_envelopes".}

\item{...}{Arguments passed to \code{\link{plot.envelope_test}}.}
}
\description{
Plot method for the class 'combined_scaled_MAD_test'.
If plot_type is "rank", then the output of the combined rank envelope test
is plotted. If plot_type is "envelope", then the combined scaled MAD envelope
is plotted.
}
