% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{plot.curve_set2d}
\alias{plot.curve_set2d}
\title{Plot method for the class 'curve_set2d'}
\usage{
\method{plot}{curve_set2d}(x, idx = 1, base_size = 11, ...)
}
\arguments{
\item{x}{An \code{curve_set2d} object}

\item{idx}{Indices of functions to plot for 2d plots.}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{...}{Additional parameters to be passed to plot and lines.}
}
\description{
Plot method for the class 'curve_set2d', i.e. two-dimensional functions
}
\examples{
data(abide_9002_23)
plot(abide_9002_23$curve_set, idx=c(1, 27))
}
