% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.envelope_test}
\alias{plot.envelope_test}
\title{Plot method for the class 'envelope_test'}
\usage{
\method{plot}{envelope_test}(x, ...)
}
\arguments{
\item{x}{an 'envelope_test' object'}

\item{...}{Additional parameters to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'envelope_test'
}
\seealso{
\code{\link{global_envelope_test}}
}
