% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.global_envelope}
\alias{plot.global_envelope}
\title{Plot method for the class 'global_envelope'}
\usage{
\method{plot}{global_envelope}(x, plot_style="basic", base_size=15, dotplot=length(x$r)<10,
                                     main, ylim, xlab, ylab, use_ggplot2, ...)
}
\arguments{
\item{x}{an 'global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{plot_style} is "basic".
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{use_ggplot2}{Logical, whether plot_style is "ggplot2" or not. Outdated, use the argument plot_style instead.}

\item{...}{Additional parameters to be passed to \code{\link{env_basic_plot}}, \code{\link{dotplot}}
(if dotplot=TRUE) or \code{\link{env_ggplot}} (if plot_style="ggplot2").}
}
\description{
Plot method for the class 'global_envelope'
}
\seealso{
\code{\link{central_region}}
}
