% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfanova.r
\name{plot.graph.fanova}
\alias{plot.graph.fanova}
\title{Plot method for the class 'graph.fanova'}
\usage{
\method{plot}{graph.fanova}(x, plot_style="ggplot2", separate_yaxes = TRUE, labels=x[['labels']], ...)
}
\arguments{
\item{x}{An 'graph.fanova' object}

\item{plot_style}{Either "basic" or "ggplot2".}

\item{separate_yaxes}{Logical (default FALSE). By default also the combined envelope plots have
a common y-axis. If TRUE, then separate y-axes are used for different parts of a combined test.}

\item{labels}{Labels for the separate plots. Ignored if separate_yaxes is FALSE.}

\item{...}{Additional parameters to be passed to \code{\link{plot.envelope_test}}.}
}
\description{
Plot method for the class 'graph.fanova'
}
