% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_set.r
\name{plot.image_set}
\alias{plot.image_set}
\title{Plot method for the class 'image_set'}
\usage{
\method{plot}{image_set}(x, idx = 1, obs = TRUE, ...)
}
\arguments{
\item{x}{an 'image_set' object}

\item{idx}{Indices of the images in the image_set to be plotted.}

\item{obs}{Logical. TRUE, then idx is understood as an index to \code{image_set$obs},
otherwise to \code{image_set$sim_m}.}

\item{main}{The title. Default exists.}

\item{col}{Colours to be passed to \code{\link[spatstat]{plot.im}}. A default exists.}

\item{...}{Additional parameters to be passed to \code{\link[spatstat]{plot.im}}.}
}
\description{
Plot method for the class 'image_set'
}
