% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgrowthmillion.r
\docType{data}
\name{popgrowthmillion}
\alias{popgrowthmillion}
\title{Population growth}
\format{
A matrix, where each row corresponds to a year and each column to a country.
Column names correspond to the countries, and row names to the years.
}
\usage{
data("popgrowthmillion")
}
\description{
Population growth
}
\details{
This dataset includes population growth, i.e. population at the end of the year divided
by population at the beginning of the year, in 134 countries in years from 1950 to 2015.
The dataset includes only countries over million inhabitants in 1950.
}
\references{
Nagy, S., I. Gijbels, and D. Hlubinka (2017). Depth-based recognition of shape outlying functions. Journal of Computational and Graphical Statistics 26 (4), 883-893.
}
\keyword{datasets}
