% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_envelopes.r
\name{print.adjusted_combined_envelope_test}
\alias{print.adjusted_combined_envelope_test}
\title{Print method for the class 'adjusted_combined_envelope_test'}
\usage{
\method{print}{adjusted_combined_envelope_test}(x, ...)
}
\arguments{
\item{x}{an 'adjusted_combined_envelope_test' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'adjusted_combined_envelope_test'
}

