% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{print.centralRegion}
\alias{print.centralRegion}
\title{Print method for the class 'centralRegion'}
\usage{
\method{print}{centralRegion}(x, ...)
}
\arguments{
\item{x}{an 'centralRegion' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'centralRegion'
}
