% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboxplot.r
\name{print.combined_fboxplot}
\alias{print.combined_fboxplot}
\title{Print method for the class 'combined_fboxplot'}
\usage{
\method{print}{combined_fboxplot}(x, ...)
}
\arguments{
\item{x}{an 'combined_fboxplot' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'combined_fboxplot'
}
