% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{print.combined_global_envelope}
\alias{print.combined_global_envelope}
\title{Print method for the class 'combined_global_envelope'}
\usage{
\method{print}{combined_global_envelope}(x, ...)
}
\arguments{
\item{x}{A 'combined_global_envelope' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'combined_global_envelope'
}
