% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes2d.r
\name{print.combined_global_envelope2d}
\alias{print.combined_global_envelope2d}
\title{Print method for the class 'combined_global_envelope2d'}
\usage{
\method{print}{combined_global_envelope2d}(x, ...)
}
\arguments{
\item{x}{an 'combined_global_envelope2d' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'combined_global_envelope2d'
}
