% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_envelope_tests.R
\name{print.combined_scaled_MAD_test}
\alias{print.combined_scaled_MAD_test}
\title{Print method for the class 'combined_scaled_MAD_test'}
\usage{
\method{print}{combined_scaled_MAD_test}(x, ...)
}
\arguments{
\item{x}{An 'combined_scaled_MAD_test' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'combined_scaled_MAD_test'
}
