% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{print.curve_set}
\alias{print.curve_set}
\title{Print method for the class 'curve_set'}
\usage{
\method{print}{curve_set}(x, ...)
}
\arguments{
\item{x}{an 'curve_set' object}

\item{...}{Passed to \code{\link[utils]{str}}.}
}
\description{
Print method for the class 'curve_set'
}
