% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation.r
\name{print.deviation_set}
\alias{print.deviation_set}
\title{Print method for the class 'deviation_set'}
\usage{
\method{print}{deviation_set}(x, ...)
}
\arguments{
\item{x}{A deviation_set object.}

\item{...}{Ignored.}
}
\description{
Print method for the class 'deviation_set'
}

