% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.envelope_test}
\alias{print.envelope_test}
\title{Print method for the class 'envelope_test'}
\usage{
\method{print}{envelope_test}(x, ...)
}
\arguments{
\item{x}{an 'envelope_test' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'envelope_test'
}

