% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{print.fboxplot}
\alias{print.fboxplot}
\title{Print method for the class 'fboxplot'}
\usage{
\method{print}{fboxplot}(x, ...)
}
\arguments{
\item{x}{an 'fboxplot' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'fboxplot'
}
