% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fclustering.r
\name{print.fclust}
\alias{print.fclust}
\title{Print method for the class 'fclust'}
\usage{
\method{print}{fclust}(x, ...)
}
\arguments{
\item{x}{A object of class 'fclust'.}

\item{...}{Ignored.}
}
\description{
Print method for the 'fclust' objects returned by \code{\link{fclustering}}.
}
