% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.r
\name{print.fdr_envelope}
\alias{print.fdr_envelope}
\title{Print method for the class 'fdr_envelope'}
\usage{
\method{print}{fdr_envelope}(x, ...)
}
\arguments{
\item{x}{An 'fdr_envelope' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'fdr_envelope'
}
