% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appl_indeptest.r
\name{print.GET_contingency}
\alias{print.GET_contingency}
\title{Print method for the class 'GET_contingency'}
\usage{
\method{print}{GET_contingency}(x, ...)
}
\arguments{
\item{x}{A 'GET_contingency' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'GET_contingency'
}
