% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{print.global_envelope}
\alias{print.global_envelope}
\title{Print method for the class 'global_envelope'}
\usage{
\method{print}{global_envelope}(x, ...)
}
\arguments{
\item{x}{A 'global_envelope' object.}

\item{...}{Ignored.}
}
\description{
Print method for the class 'global_envelope'
}
