% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfanova.r
\name{print.graph.fanova}
\alias{print.graph.fanova}
\title{Print method for the class 'graph.fanova'}
\usage{
\method{print}{graph.fanova}(x, ...)
}
\arguments{
\item{x}{an 'graph.fanova' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'graph.fanova'
}
