% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_set.r
\name{print.image_set}
\alias{print.image_set}
\title{Print method for the class 'image_set'}
\usage{
\method{print}{image_set}(x, ...)
}
\arguments{
\item{x}{an 'image_set' object}

\item{...}{Passed to \code{\link[utils]{str}}.}
}
\description{
Print method for the class 'image_set'
}
