% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residual.r
\name{residual}
\alias{residual}
\title{Subtract S_{H_0} from the summary functions.}
\usage{
residual(curve_set, use_theo = TRUE)
}
\arguments{
\item{curve_set}{A curve_set or an \code{\link[spatstat]{envelope}}
object. If an envelope object is given, it must contain the summary
functions from the simulated patterns which can be achieved by setting
savefuns = TRUE when calling envelope().}

\item{use_theo}{Whether to use the theoretical summary function or the
mean of the simulations.}
}
\value{
A curve set object containing residual summary functions. theo is
  no longer included.
}
\description{
Subtract S_{H_0} from the summary functions.
}

