\name{scaled_deviation_test}
\alias{scaled_deviation_test}
\title{Scale the curve_set and perform a deviation test.}
\usage{
  scaled_deviation_test(curve_set, r_min = NULL,
    r_max = NULL, use_theo = TRUE, scaling = "env",
    measure = "max")
}
\description{
  Scale the curve_set and perform a deviation test.
}

