\docType{package}
\name{spptest}
\alias{spptest}
\alias{spptest-package}
\title{Spatial point process testing}
\description{
  spptest provides envelope and deviation tests for spatial
  point processes. The main motivation for a new package
  are the power-correcting scalings for deviation tests,
  new deviation measures and new envelope tests.
}
\references{
  In preparation.
}

