% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.r
\name{st_scaling}
\alias{st_scaling}
\title{Studentised scaling}
\usage{
st_scaling(curve_set)
}
\arguments{
\item{curve_set}{A curve_set or an \code{\link[spatstat]{envelope}}
object. If an envelope object is given, it must contain the summary
functions from the simulated patterns which can be achieved by setting
savefuns = TRUE when calling envelope().}
}
\value{
A scaled curve_set.
}
\description{
Scale with the standard deviation.
}

