% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{subset.curve_set}
\alias{subset.curve_set}
\title{Subsetting curve sets}
\usage{
\method{subset}{curve_set}(x, subset, ...)
}
\arguments{
\item{x}{A \code{curve_set} object.}

\item{subset}{A logical expression indicating curves to keep.}

\item{...}{Ignored.}
}
\description{
Return subsets of curve sets which meet conditions.
}
