% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ThemePlain}
\alias{ThemePlain}
\title{A plain ggplot2 theme.}
\usage{
ThemePlain(base_size = 15, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A plain ggplot2 theme.
}

