\name{weigh_both_sides}
\alias{weigh_both_sides}
\title{Weigh a matrix or a vector with two different coeffs depending on which
side of middle curve each element is.}
\usage{
  weigh_both_sides(x, upper_coeff, lower_coeff)
}
\arguments{
  \item{x}{The matrix}

  \item{upper_coeff}{Upper coefficient.}

  \item{lower_coeff}{Lower coefficient.}
}
\description{
  Used by \code{\link{qdir_scaling}}.
}

