% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.r
\name{weigh_curves}
\alias{weigh_curves}
\title{Multiply by a coefficient.}
\usage{
weigh_curves(curve_set, coeff)
}
\arguments{
\item{curve_set}{A curve_set or an \code{\link[spatstat]{envelope}}
object. If an envelope object is given, it must contain the summary
functions from the simulated patterns which can be achieved by setting
savefuns = TRUE when calling envelope().}

\item{coeff}{The coefficient vector, often of the length of one curve.}
}
\description{
Multiply by a coefficient.
}

