\name{check_curve_set_content}
\alias{check_curve_set_content}
\title{Check the content validity of a potential curve_set object.}
\usage{
  check_curve_set_content(curve_set)
}
\arguments{
  \item{curve_set}{A curve_set object to be checked.}
}
\description{
  Check the content validity of a potential curve_set
  object.
}

