% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{check_residualness}
\alias{check_residualness}
\title{Check that the curve_set object portrays residual curves.}
\usage{
check_residualness(curve_set)
}
\arguments{
\item{curve_set}{A 'curve_set' object}
}
\description{
Check that the curve_set object portrays residual curves.
}
