% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{create_curve_set}
\alias{create_curve_set}
\title{Create a curve set out of a list in the right form.}
\usage{
create_curve_set(curve_set, ...)
}
\arguments{
\item{curve_set}{A list containing elements r, obs, and optionally
the elements sim_m and theo.
\code{r} must be a vector describing the radius vector, the argument values where
functions have been observed (or simulated).
\code{obs} must be either 1) a vector containing the data function, in which case
\code{obs} must have same length as \code{r}, or
2) a matrix containing the s data functions, in which case it is assumed that
each column corresponds to a data function, and the number of rows must match the length or r.
If \code{obs} is a vector, \code{sim_m} must be a matrix containing the simulated functions.
Each column is assumed to correspond to a function, and the number of rows must match the
length of r. If \code{obs} is a matrix, \code{sim_m} is ignored.
If included, \code{theo} corresponds to the theoretical function (e.g., under the null hypothesis).
If present, its length must match the length of \code{r}.}

\item{...}{Do not use. (For internal use only.)}
}
\value{
The given list adorned with the proper class name.
}
\description{
Create a curve set out of a list in the right form.
}
