% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation.r
\name{create_deviation_set}
\alias{create_deviation_set}
\title{Create a deviation set out of a list in the right form.}
\usage{
create_deviation_set(deviation_set)
}
\arguments{
\item{deviation_set}{The list to be turned into a deviation_set object.}
}
\description{
Create a deviation set out of a list in the right form.
}

