% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p-value.r
\name{estimate_p_value.deviation_set}
\alias{estimate_p_value.deviation_set}
\title{Estimate p-value.}
\usage{
\method{estimate_p_value}{deviation_set}(deviation_set, ...)
}
\arguments{
\item{deviation_set}{A deviation_set object. Use deviation() to get one.}

\item{...}{Additional parameters to be passed to \code{\link{estimate_p_value.default}}.}
}
\value{
The p-value.
}
\description{
Check \code{\link{estimate_p_value.default}} for details.
}

