% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate_p_value}
\alias{estimate_p_value}
\title{Estimate p-value.}
\usage{
estimate_p_value(x, ...)
}
\arguments{
\item{x}{The first argument.}

\item{...}{Additional arguments.}
}
\description{
Estimate p-value.
}
\seealso{
\code{\link{estimate_p_value.default}}
}

