% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfanova.r
\name{graph.fanova}
\alias{graph.fanova}
\title{Graphical functional ANOVA}
\usage{
graph.fanova(nsim, x, groups, summaryfun, alpha = 0.05, n.aver = 1L,
  mirror = FALSE)
}
\arguments{
\item{nsim}{The number of random permutations.}

\item{x}{The original data (an array of functions). Typically a matrix or a data frame,
also \code{\link[fda.usc]{fdata}} objects allowed.}

\item{groups}{The original froups (a factor vector representing the assignment to groups).}

\item{summaryfun}{Possible values are "means", "studmeans", "contrasts", "studcontrasts".}

\item{alpha}{The significance level of the test.}

\item{mirror}{The complement of the argument circular of \code{\link[stats]{filter}}.}

\item{n.iter}{If summaryfun is either "studmeans" or "studcontrasts", there is a possibility to
use variances smoothed by appying moving average to the estimated variance. n.iter determines
how many values on each side do contribute (incl. value itself).}
}
\description{
One-way ANOVA tests for functional data with graphical interpretation
}
\examples{
# Generate an array of 1999 plain means by permutation.
data(rimov)
groups <- factor(c(rep(1, times=12), rep(2, times=12), rep(3, times=12)))

nsim <- 2499 # for exploring

res <- graph.fanova(nsim=nsim, x=rimov, groups=groups, summaryfun="means")
plot(res)
res2 <- graph.fanova(nsim=nsim, x=rimov, groups=groups, summaryfun="contrasts")
plot(res2)
}
\references{
Mrkvička, T., Hahn, U. and Myllymäki, M. (2016)
A one-way ANOVA test for functional data with graphical interpretation.
arXiv:1612.03608 [stat.ME] (http://arxiv.org/abs/1612.03608)
}
