% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.combined_fboxplot}
\alias{plot.combined_fboxplot}
\title{Plot method for the class 'combined_fboxplot'.}
\usage{
\method{plot}{combined_fboxplot}(x, max_ncols_of_plots = 2, main, curve_sets, ...)
}
\arguments{
\item{x}{an 'combined_fboxplot' object}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2.}

\item{main}{A list of titles for the separate plots. The length should be the length of x.}

\item{curve_sets}{A list of curve_sets, most typically those from which the combined functional
boxplot has been contructed.}

\item{...}{Additional parameters to be passed to \code{\link{plot.fboxplot}}.}
}
\description{
Current method is just to plot the components of the combined functional boxplot
one after each other.
}
