% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.combined_fboxplot}
\alias{plot.combined_fboxplot}
\title{Plot method for the class 'combined_fboxplot'}
\usage{
\method{plot}{combined_fboxplot}(x, level = 1, outliers = TRUE, bp.col = 2, cr.col = 1, ...)
}
\arguments{
\item{x}{an 'combined_fboxplot' object}

\item{level}{1 or 2. In the case of two-step combined tests (with several test functions), two different plots are available:
1 for plotting the combined global envelopes (default and most often wanted) or
2 for plotting the second level test result.}

\item{outliers}{Logical. If TRUE, then the functions outside the functional boxplot are drawn.}

\item{bp.col}{The color for the boxplot bounds. Default 2 (red).}

\item{cr.col}{The color for the central region bounds.}

\item{...}{Additional arguments to be passed to \code{\link{plot.combined_global_envelope}}.}
}
\description{
Plot method for the class 'combined_fboxplot'
}
