% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.combined_global_envelope}
\alias{plot.combined_global_envelope}
\title{Plot method for the class 'combined_global_envelope'}
\usage{
\method{plot}{combined_global_envelope}(x, plot_style="basic", level = 1, max_ncols_of_plots=2, ...)
}
\arguments{
\item{x}{an 'combined_global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{level}{1 for plotting the combined global envelopes or
2 for plotting the second level ERL test result.}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2.}

\item{...}{Additional parameters to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'combined_global_envelope'
}
