% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.combined_global_envelope}
\alias{plot.combined_global_envelope}
\title{Plot method for the class 'combined_global_envelope'}
\usage{
\method{plot}{combined_global_envelope}(x, plot_style = c("ggplot2", "fv", "basic"),
   main, ylim, xlab, ylab,
   color_outside = TRUE, env.col = 1, base_size = 15,
   labels = NULL, add = FALSE, digits=3,
   level = 1, max_ncols_of_plots = 2, nticks = 5,
   legend = TRUE, ...)
}
\arguments{
\item{x}{An 'combined_global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used. Relevant only for \code{plot_style = "basic"}.}

\item{env.col}{The color for the envelope lines (or dotplot arrows). Default 1 (black).}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{labels}{A character vector of suitable length.
If \code{dotplot = TRUE} (for the level 2 test), then labels for the tests at x-axis.
Otherwise labels for the separate plots.}

\item{add}{Whether to add the plot to an existing plot (TRUE) or to draw a new plot (FALSE).
Not available for \code{plot_style = "ggplot2"}.}

\item{digits}{The number of digits used for printing the p-value or p-interval in the main,
if using the default main.}

\item{level}{1 or 2. In the case of two-step combined tests (with several test functions), two different plots are available:
1 for plotting the combined global envelopes (default and most often wanted) or
2 for plotting the second level test result.}

\item{max_ncols_of_plots}{The maximum number of columns for the figures. Default 2.
(Relates to the number of curve_sets that have been combined.)}

\item{nticks}{The number of ticks on the xaxis.}

\item{legend}{Logical. If FALSE, then the legend is removed from the "ggplot2" style plot.}

\item{...}{Additional parameters to be passed to \code{\link{plot}} or \code{\link{lines}}.}
}
\description{
Plot method for the class 'combined_global_envelope'
}
\seealso{
\code{\link{central_region}}
}
