% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{plot.curve_set}
\alias{plot.curve_set}
\title{Plot method for the class 'curve_set'}
\usage{
\method{plot}{curve_set}(x, ylim, xlab = "r", ylab = "obs",
  col_obs = 1, col_sim = grDevices::grey(0.7), ...)
}
\arguments{
\item{x}{An \code{curve_set} object}

\item{ylim}{The y limits of the plot with the default being the minimum and maximum over all curves.}

\item{xlab}{The label for the x-axis. Default "r".}

\item{ylab}{The label for the y-axis. Default "obs".}

\item{col_obs}{Color for 'obs' in the argument \code{x}.}

\item{col_sim}{Color for 'sim_m' in the argument \code{x}.}

\item{...}{Additional parameters to be passed to plot and lines.}
}
\description{
Plot method for the class 'curve_set'
}
