\name{plot.curve_set}
\alias{plot.curve_set}
\title{Plot method for the class 'curve_set'}
\usage{
  \method{plot}{curve_set}(x, ...)
}
\arguments{
  \item{x}{an 'curve_set' object}

  \item{...}{Additional parameters to be passed to plot and
  lines.}
}
\description{
  Plot method for the class 'curve_set'
}

