% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-set.r
\name{plot.curve_set}
\alias{plot.curve_set}
\title{Plot method for the class 'curve_set'}
\usage{
\method{plot}{curve_set}(x, ylim, ...)
}
\arguments{
\item{x}{an 'curve_set' object}

\item{ylim}{The y limits of the plot with the default being the minimum and maximum over all curves.}

\item{...}{Additional parameters to be passed to plot and lines.}
}
\description{
Plot method for the class 'curve_set'
}
