% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.envelope_test}
\alias{plot.envelope_test}
\title{Plot method for the class 'envelope_test'}
\usage{
\method{plot}{envelope_test}(x, use_ggplot2=FALSE, base_size=15, dotplot=length(x$r)<10,
                                     main, ylim, xlab, ylab, ...)
}
\arguments{
\item{x}{an 'envelope_test' object}

\item{use_ggplot2}{TRUE/FALSE, If TRUE, then a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.}

\item{base_size}{Base font size, to be passed to theme style when \code{use_ggplot2 = TRUE}.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{use_ggplot2} is FALSE.
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists. If the test is a combined
test and the argument separate_yaxes is TRUE, then a vector containing xlabs for each part of the
test can be given.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists. If the test is a combined
test and the argument separate_yaxes is TRUE, then a vector containing ylabs for each part of the
test can be given.}

\item{...}{Additional parameters to be passed to \code{\link{env_basic_plot}}, \code{\link{dotplot}}
(if dotplot=TRUE) or \code{\link{env_ggplot}} (if use_ggplot2=TRUE).}
}
\description{
Plot method for the class 'envelope_test'
}
\seealso{
\code{\link{rank_envelope}}, \code{\link{st_envelope}}, \code{\link{qdir_envelope}}
}

