% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.envelope_test}
\alias{plot.envelope_test}
\title{Plot method for the class 'envelope_test'}
\usage{
\method{plot}{envelope_test}(x, use_ggplot2=FALSE, base_size=15, dotplot=length(x$r)<10,
                                     color_outside=TRUE, main, ylim, xlab, ylab, ...)
}
\arguments{
\item{x}{an 'envelope_test' object}

\item{use_ggplot2}{TRUE/FALSE, If TRUE, then a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.}

\item{base_size}{Base font size, to be passed to theme style when \code{use_ggplot2 = TRUE}.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{use_ggplot2} is FALSE.
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{color_outside}{Logical. Whether to color the places where the data function goes outside the envelope.
Currently red color is used and coloring is only used if \code{use_ggplot2} is FALSE.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{...}{Additional parameters to be passed to plot (if use_ggplot2=FALSE).}
}
\description{
Plot method for the class 'envelope_test'
}
\seealso{
\code{\link{rank_envelope}}, \code{\link{st_envelope}}, \code{\link{qdir_envelope}}
}

