% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboxplot.r
\name{plot.fboxplot}
\alias{plot.fboxplot}
\title{Plot method for the class 'fboxplot'}
\usage{
\method{plot}{fboxplot}(x, main, xlab, ylab, digits = 3, legend = TRUE, plot_outliers = TRUE, ...)
}
\arguments{
\item{x}{an 'fboxplot' object}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{digits}{The number of digits used for printing the p-value or p-interval in the main,
if using the default main.}

\item{legend}{Logical. See \code{plot_outliers.}}

\item{plot_outliers}{Logical. If TRUE, then the functions outside the functional boxplot are drawn,
with legend if \code{legend} is TRUE.}

\item{...}{Additional arguments to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'fboxplot'
}
