% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.fboxplot}
\alias{plot.fboxplot}
\title{Plot method for the class 'fboxplot'}
\usage{
\method{plot}{fboxplot}(x, plot_style = "basic", curve_set, dotplot = length(x$r)<10, ylim, ...)
}
\arguments{
\item{x}{an 'fboxplot' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{curve_set}{A curve_set of functions, most likely the one from which the functional boxplot
has been created. If given, then the functions outside the functional boxplot are drawn.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{plot_style} is "basic".
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{...}{Additional arguments to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'fboxplot'
}
