% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.global_envelope}
\alias{plot.global_envelope}
\title{Plot method for the class 'global_envelope'}
\usage{
\method{plot}{global_envelope}(x, plot_style = c("basic", "fv", "ggplot2"),
   dotplot = length(x$r)<10,
   main, ylim, xlab, ylab, use_ggplot2,
   color_outside = TRUE, env.col = 1, base_size = 15,
   labels = NULL, add = FALSE, digits=3,
   level = 1, separate_yaxes = TRUE, max_ncols_of_plots = 2,
   nticks = 5, ...)
}
\arguments{
\item{x}{An 'global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{plot_style} is "basic".
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{use_ggplot2}{Logical, whether plot_style is "ggplot2" or not. Outdated, use the argument plot_style instead.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used. Relevant only for \code{plot_style = "basic"}.}

\item{env.col}{The color for the envelope lines (or dotplot arrows). Default 1 (black).}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{labels}{A character vector of suitable length.
If \code{dotplot = TRUE}, then labels for the tests at x-axis.
Labels for the separate plots, ignored if separate_yaxes is FALSE.}

\item{add}{Whether to add the plot to an existing plot (TRUE) or to draw a new plot (FALSE).
Not available for \code{plot_style = "ggplot2"}.}

\item{digits}{The number of digits used for printing the p-value or p-interval in the main,
if using the default main.}

\item{level}{1 or 2. In the case of combined tests (with several test functions), two different plots are available:
1 for plotting the combined global envelopes (default and most often wanted) or
2 for plotting the second level test result.}

\item{separate_yaxes}{Logical (default FALSE). If TRUE, then separate y-axes are used for
different parts of a combined test. If FALSE, then the components are displayed one next to each other
with a joint y-axis.}

\item{max_ncols_of_plots}{If separate_yaxes is TRUE, then max_ncols_of_plots gives the maximum
number of columns for figures. Default 2. Relevant only for combined tests with several test functions.}

\item{nticks}{The number of ticks on the xaxis. Relevant only for combined tests with
\code{separate_yaxes = TRUE}.}

\item{...}{Additional parameters to be passed to \code{\link{plot}} or \code{\link{lines}}.}
}
\description{
Plot method for the class 'global_envelope'
}
\seealso{
\code{\link{central_region}}
}
