% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes.r
\name{plot.global_envelope}
\alias{plot.global_envelope}
\title{Plot method for the class 'global_envelope'}
\usage{
\method{plot}{global_envelope}(
  x,
  plot_style = c("ggplot2", "fv", "basic"),
  dotplot = length(x$r) < 10,
  main,
  ylim,
  xlab,
  ylab,
  coord = NULL,
  color_outside = TRUE,
  env.col = 1,
  base_size = 11,
  labels = NULL,
  add = FALSE,
  digits = 3,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{An 'global_envelope' object}

\item{plot_style}{One of the following "basic", "fv" or "ggplot2".
The option "basic" (default) offers a very basic global envelope plot.
The option "fv" utilizes the plot routines of the function value table \code{\link[spatstat]{fv.object}}.
For "ggplot2", a plot with a coloured envelope ribbon is provided. Requires R library ggplot2.
The option "fv" is currently only available for tests with one test function, whereas the other true allow
also tests with several tests functions.}

\item{dotplot}{Logical. If TRUE, then instead of envelopes a dot plot is done.
Suitable for low dimensional test vectors. Only applicable if \code{plot_style} is "basic".
Default: TRUE if the dimension is less than 10, FALSE otherwise.}

\item{main}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylim}{See \code{\link{plot.default}}. A sensible default exists.}

\item{xlab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{ylab}{See \code{\link{plot.default}}. A sensible default exists.}

\item{coord}{A data frame of the spatial coordinates where the data have been observed.
\code{nrow(coord)} should match the length of \code{x$r}, and the names of the columns
should be either "x", "y", "width", "height" or "xmin", "ymin", "xmax", "ymax".
Here x and y should give the (center) coordinates of the observed data,
width and height give the size of the pixel places at (x,y) (most often constants).
The xmin, ymin, xmax, ymax give an alternative way to specify the pixels where the data have
been observed, namely the corner locations of each pixel.
If coord is provided, then a two-dimensional plot is made instead of 1-dimensional default
plot.}

\item{color_outside}{Logical. Whether to color the places where the data function goes
outside the envelope. Currently red color is used. Relevant only for \code{plot_style = "basic"}.}

\item{env.col}{The color for the envelope lines (or dotplot arrows). Default 1 (black).}

\item{base_size}{Base font size, to be passed to theme style when \code{plot_style = "ggplot2"}.}

\item{labels}{A character vector of suitable length.
If \code{dotplot = TRUE}, then labels for the tests at x-axis.}

\item{add}{Whether to add the plot to an existing plot (TRUE) or to draw a new plot (FALSE).
Not available for \code{plot_style = "ggplot2"}.}

\item{digits}{The number of digits used for printing the p-value or p-interval in the main,
if using the default main.}

\item{legend}{Logical. If FALSE, then the legend is removed from the "ggplot2" style plot.}

\item{...}{Additional parameters to be passed to \code{\link{plot}} or \code{\link{lines}}.}
}
\description{
Plot method for the class 'global_envelope'
}
\seealso{
\code{\link{central_region}}
}
