% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopes2d.r
\name{plot.global_envelope_2d}
\alias{plot.global_envelope_2d}
\title{Plot method for the class 'global_envelope_2d'}
\usage{
\method{plot}{global_envelope_2d}(x, col, col.sign = c(255, 0, 0), transparency = 85, main, ...)
}
\arguments{
\item{x}{an 'global_envelope_2d' object}

\item{col}{As \code{col} in \code{\link[spatstat]{plot.im}}.
If not given, a \code{\link[spatstat]{colourmap}} of 255 grey values between the minimum and maximum
of the observed function (if present) and lower and upper boundary of the global envelope is used.
The same \code{col} is used for each produced plot to make comparisons easier.}

\item{col.sign}{A vector of length 3 giving the color for the significant regions.
The elements should correspond to red, green, blue of \code{\link[grDevices]{rgb}}.
Default to}

\item{transparency}{A number between 0 and 255 (default 85, 33% transparency).
Corresponds to alpha of \code{\link[grDevices]{rgb}}. Used in plotting the significant regions.}

\item{main}{A vector of length 5 giving titles for the 4 (if there is an observed function) or
5 plots. A default exists, but this argument allows modifying it.
The default is \code{c("observed", "lo envelope", "hi envelope", "Significance: below (red)",
"Significance: above (red)")}.}

\item{...}{Additional parameters to be passed to \code{\link[spatstat]{plot.im}}.}
}
\description{
Plot method for the class 'global_envelope_2d'
}
