% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfanova.r
\name{plot.graph.fanova}
\alias{plot.graph.fanova}
\title{Plot method for the class 'graph.fanova'}
\usage{
\method{plot}{graph.fanova}(x, plot_style="ggplot2", separate_yaxes = TRUE, labels=x[['labels']], ...)
}
\arguments{
\item{x}{An 'graph.fanova' object}

\item{plot_style}{Either "basic" or "ggplot2".}

\item{separate_yaxes}{Logical (default FALSE). If TRUE, then separate y-axes are used for
different parts of a combined test. If FALSE, then the components are displayed one next to each other
with a joint y-axis.}

\item{labels}{Labels for the separate plots (for \code{plot_style = "ggplot2"}).
Ignored if separate_yaxes is FALSE. Or, for \code{dotplot=TRUE}, labels for the tests at x-axis.}

\item{...}{Additional parameters to be passed to \code{\link{plot.global_envelope}}.}
}
\description{
Plot method for the class 'graph.fanova'
}
