% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_set.r
\name{plot.image_set}
\alias{plot.image_set}
\title{Plot method for the class 'image_set'}
\usage{
\method{plot}{image_set}(
  x,
  idx = 1,
  obs = TRUE,
  plot_style = c("ggplot2", "basic"),
  main,
  col,
  max_ncols_of_plots = 4,
  ...
)
}
\arguments{
\item{x}{an 'image_set' object}

\item{idx}{Indices of the images in the image_set to be plotted.}

\item{obs}{Logical. TRUE, then idx is understood as an index to \code{image_set$obs},
otherwise to \code{image_set$sim_m}.}

\item{plot_style}{Either "ggplot2" or "basic". (Similar to the argument in \code{\link{plot.global_envelope}}.)}

\item{main}{The title. Default exists.}

\item{col}{Colours to be passed to \code{\link[spatstat]{plot.im}} if
\code{plot_style = "basic"}. A default exists.}

\item{max_ncols_of_plots}{The maximum number of columns for the figures. Default 4.}

\item{...}{Additional parameters to be passed to \code{\link[spatstat]{plot.im}}.}
}
\description{
Plot method for the class 'image_set'
}
