% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation_test.r
\name{print.deviation_test}
\alias{print.deviation_test}
\title{Print method for the class 'deviation_test'}
\usage{
\method{print}{deviation_test}(x, ...)
}
\arguments{
\item{x}{an 'deviation_test' object}

\item{...}{Ignored.}
}
\description{
Print method for the class 'deviation_test'
}
