% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saplings.r
\docType{data}
\name{saplings}
\alias{saplings}
\title{Saplings data set}
\format{An object of class \code{"ppp"}
representing the point pattern of tree locations.

See \code{\link[spatstat]{ppp.object}} for details of the format.}
\usage{
data(saplings)
}
\description{
A pattern of small trees (height <= 15 m) originating from an uneven aged multi-species
broadleaf nonmanaged forest in Kaluzhskie Zaseki, Russia.
}
\references{
Grabarnik, P. and Chiu, S. N. (2002) Goodness-of-fit test for complete spatial randomness against
mixtures of regular and clustered spatial point processes. \emph{Biometrika}, \bold{89}, 411–421.

van Lieshout, M.-C. (2010) Spatial point process theory. In Handbook of Spatial Statistics (eds. A. E.
Gelfand, P. J. Diggle, M. Fuentes and P. Guttorp), Handbooks of Modern Statistical Methods. Boca
Raton: CRC Press.
}
\keyword{datasets}
\keyword{spatial}

